+++ src/Console/Command/UserUpdateCommand.php
@@ -7,21 +7,22 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
 
-use Inet\SugarCRM\Exception\SugarException;
+use Inet\SugarCRM\Exception\BeanNotFoundException;
 use Inet\SugarCRM\UsersManager;
 
-class UsersUpdateCommand extends AbstractConfigOptionCommand
+use SugarCli\Console\ExitCode;
+
+class UserUpdateCommand extends AbstractConfigOptionCommand
 {
     protected $fields_mapping = array(
         'first-name' => 'first_name',
         'last-name' => 'last_name',
-        'email' => 'email1',
     );
 
     protected function configure()
     {
-        $this->setName('users:update')
-            ->setAliases(array('users:create'))
+        $this->setName('user:update')
+            ->setAliases(array('user:create'))
             ->setDescription('Create or update a SugarCRM user.')
             ->addConfigOptionMapping('path', 'sugarcrm.path')
             ->addArgument(
@@ -54,19 +55,13 @@ class UsersUpdateCommand extends AbstractConfigOptionCommand
                 'Password of the user.'
             )
             ->addOption(
-                'email',
-                'e',
-                InputOption::VALUE_REQUIRED,
-                'Email address of the user.'
-            )
-            ->addOption(
                 'admin',
                 'a',
                 InputOption::VALUE_REQUIRED,
                 'Make the user administrator. <comment>[yes/no]</comment>'
             )
             ->addOption(
-                'activated',
+                'active',
                 'A',
                 InputOption::VALUE_REQUIRED,
                 'Make the user active. <comment>[yes/no]</comment>'
@@ -75,11 +70,16 @@ class UsersUpdateCommand extends AbstractConfigOptionCommand
 
     protected function isCreate(InputInterface $input)
     {
-        list($prefix, $cmd) = $input->getFirstArgument();
-        if (substr_compare("create", $cmd, 0) === 0) {
+        list($prefix, $cmd) = explode(':', $input->getFirstArgument(), 2);
+        if (substr_compare("create", $cmd, 0, strlen($cmd)) === 0) {
             return true;
         }
         return $input->getOption('create');
+    }
+
+    public function getBoolean($value)
+    {
+        return filter_var($value, FILTER_VALIDATE_BOOLEAN);
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -89,7 +89,7 @@ class UsersUpdateCommand extends AbstractConfigOptionCommand
         $user_name = $input->getArgument('username');
         $password = $input->getOption('password');
         $admin = $input->getOption('admin');
-        $activated = $input->getOption('activated');
+        $active = $input->getOption('active');
 
         $additionnal_fields = array();
         foreach ($this->fields_mapping as $option => $field_name) {
@@ -102,22 +102,24 @@ class UsersUpdateCommand extends AbstractConfigOptionCommand
             $um = new UsersManager($this->getService('sugarcrm.entrypoint'));
             if ($this->isCreate($input)) {
                 $um->createUser($user_name, $additionnal_fields);
+                if (is_null($activated)) {
+                    $um->activate($user_name);
+                }
             } else {
                 $um->updateUser($user_name, $additionnal_fields);
             }
             if (!is_null($admin)) {
-                $um->setAdmin($user_name, $admin);
+                $um->setAdmin($user_name, $this->getBoolean($admin));
             }
-            if (!is_null($activated)) {
-                $um->setActive($activated);
+            if (!is_null($active)) {
+                $um->setActive($user_name, $this->getBoolean($active));
             }
             if (!is_null($password)) {
                 $um->setPassword($user_name, $password);
             }
-        } catch (SugarException $e) {
-            $logger->error('An error occured with the sugar application.');
-            $logger->error($e->getMessage());
-            return ExitCode::EXIT_UNKNOWN_SUGAR_ERROR;
+        } catch (BeanNotFoundException $e) {
+            $logger->error("User '{$user_name}' doesn't exists on the SugarCRM located at '{$path}'.");
+            return ExitCode::EXIT_USER_NOT_FOUND;
         }
     }
 }